pyproject.toml
[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | _build
    | build
    | dist
  )/
  | requirements.txt
                     # the root of the project
)
'''

[tool.poetry]
name = "photoupload"
version = "1.0.1"
description = ""
authors = ["EdLeafe <ed@leafe.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.0.1"
boto = "^2.49.0"
PyMySQL = "^1.0.2"
click = "^8.0.3"
pyexif = {path = "../pyexif"}
requests = "^2.27.1"
common = {path = "../common", develop = true}

[tool.poetry.dev-dependencies]
black = "^22.1.0"
ipython = "^8.0.1"
pytest-pudb = "^0.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
